ci: Improve versioning by using git tags
The versioning logic in the GitHub Actions workflow has been updated. It now calculates the patch version based on the last git tag instead of using the `github.run_number`.

If the major or minor version in `gradle.properties` has changed since the last tag, the patch version is reset to 1. Otherwise, it's incremented from the last tag's patch number. The `versionCode` is now set to this calculated patch number.